+++ pendulum/parsing/parser.py
@@ -3,6 +3,7 @@
 import re
 import copy
 
+from datetime import datetime
 from dateutil import parser
 
 from .exceptions import ParserError
@@ -15,8 +16,8 @@ class Parser(object):
         '^'
         '(\d{4})'  # Year
         '('
-        '    ((?:-|/)?(\d{1,2}))?'  # Month (optional)
-        '    ((?:-|/)?(\d{1,2}))?'  # Day (optional)
+        '    ((?:-|/)?(\d{2,3}))?'  # Month (optional)
+        '    ((?:-|/)?(\d{2}))?'  # Day (optional)
         ')?'
 
         # Time (Optional)
@@ -30,11 +31,22 @@ class Parser(object):
         '    )?'
         # Timezone offset
         '    ('
-        '        (-|\+)\d{2}:?\d{2}|Z'  # Offset (+HH:mm or +HHmm or Z)
+        '        (-|\+)\d{2}:?(?:\d{2})?|Z'  # Offset (+HH:mm or +HHmm or +HH or Z)
         '    )?'
         ')?'
         '$',
         re.VERBOSE
+    )
+
+    ISO8601_WEEK = re.compile(
+        '^'
+        '(\d{4})'  # Year
+        '-?'  # Separator (optional)
+        'W'  # W separator
+        '(\d{2})'  # Week number
+        '-?'  # Separator (optional)
+        '(\d)?'  # Weekday (optional)
+        '$'
     )
 
     DEFAULT_OPTIONS = {
@@ -64,6 +76,11 @@ class Parser(object):
                 day = 1
             else:
                 if m.group(4) and m.group(6):
+                    if len(m.group(4)) == 3:
+                        # Should not happen
+                        # Ordinal day with extra day set
+                        raise ParserError('Invalid date string: {}'.format(text))
+
                     # Month and day
                     if self._options['day_first']:
                         month = int(m.group(6))
@@ -73,8 +90,17 @@ class Parser(object):
                         day = int(m.group(6))
                 else:
                     # Only month
-                    month = int(m.group(4) or m.group(6))
-                    day = 1
+                    if m.group(4) and len(m.group(4)) == 3:
+                        # Ordinal day
+                        dt = datetime.strptime(
+                            '{}-{}'.format(year, m.group(4)),
+                            '%Y-%j'
+                        )
+                        month = dt.month
+                        day = dt.day
+                    else:
+                        month = int(m.group(4) or m.group(6))
+                        day = 1
 
             parsed = {
                 'year': year,
@@ -111,6 +137,9 @@ class Parser(object):
                     negative = True if tz.startswith('-') else False
                     tz = tz[1:]
                     if ':' not in tz:
+                        if len(tz) == 2:
+                            tz = '{}00'.format(tz)
+
                         off_hour = tz[0:2]
                         off_minute = tz[2:4]
                     else:
@@ -125,6 +154,34 @@ class Parser(object):
 
             return parsed
 
+    def parse_8601_week(self, text):
+        m = self.ISO8601_WEEK.match(text)
+
+        if not m:
+            return {}
+
+        year = m.group(1)
+        week = m.group(2)
+        weekday = m.group(3)
+        if not weekday:
+            weekday = '1'
+
+        fmt = '%YW%W%w'
+        string = '{}W{}{}'.format(year, week, weekday)
+
+        dt = datetime.strptime(string, fmt)
+
+        return {
+            'year': dt.year,
+            'month': dt.month,
+            'day': dt.day,
+            'hour': dt.hour,
+            'minute': dt.minute,
+            'second': dt.second,
+            'subsecond': dt.microsecond * 1000,
+            'offset': dt.utcoffset().total_seconds() if dt.tzinfo else None,
+        }
+
     def parse(self, text):
         """
         Parses a string with the given options.
@@ -134,6 +191,11 @@ class Parser(object):
 
         :rtype: dict
         """
+        # ISO8601 week notation
+        parsed = self.parse_8601_week(text)
+        if parsed:
+            return parsed
+
         parsed = self.parse_common(text)
         if parsed:
             return parsed
@@ -150,6 +212,6 @@ class Parser(object):
             'hour': dt.hour,
             'minute': dt.minute,
             'second': dt.second,
-            'subsecond': dt.microsecond,
+            'subsecond': dt.microsecond * 1000,
             'offset': dt.utcoffset().total_seconds() if dt.tzinfo else None,
         }