+++ tests/parsing_test/test_parser.py
@@ -147,12 +147,12 @@ class ParserTest(AbstractTestCase):
         self.assertEqual(123456789, parsed['subsecond'])
         self.assertEqual(19800, parsed['offset'])
 
-    def test_iso_8601(self):
-        text = '201610'
+    def test_iso_8601_date(self):
+        text = '2012'
 
         parsed = Parser().parse(text)
-        self.assertEqual(2016, parsed['year'])
-        self.assertEqual(10, parsed['month'])
+        self.assertEqual(2012, parsed['year'])
+        self.assertEqual(1, parsed['month'])
         self.assertEqual(1, parsed['day'])
         self.assertEqual(0, parsed['hour'])
         self.assertEqual(0, parsed['minute'])
@@ -160,6 +160,55 @@ class ParserTest(AbstractTestCase):
         self.assertEqual(0, parsed['subsecond'])
         self.assertEqual(None, parsed['offset'])
 
+        text = '2012-05-03'
+
+        parsed = Parser().parse(text)
+        self.assertEqual(2012, parsed['year'])
+        self.assertEqual(5, parsed['month'])
+        self.assertEqual(3, parsed['day'])
+        self.assertEqual(0, parsed['hour'])
+        self.assertEqual(0, parsed['minute'])
+        self.assertEqual(0, parsed['second'])
+        self.assertEqual(0, parsed['subsecond'])
+        self.assertEqual(None, parsed['offset'])
+
+        text = '20120503'
+
+        parsed = Parser().parse(text)
+        self.assertEqual(2012, parsed['year'])
+        self.assertEqual(5, parsed['month'])
+        self.assertEqual(3, parsed['day'])
+        self.assertEqual(0, parsed['hour'])
+        self.assertEqual(0, parsed['minute'])
+        self.assertEqual(0, parsed['second'])
+        self.assertEqual(0, parsed['subsecond'])
+        self.assertEqual(None, parsed['offset'])
+
+        text = '2012-05'
+
+        parsed = Parser().parse(text)
+        self.assertEqual(2012, parsed['year'])
+        self.assertEqual(5, parsed['month'])
+        self.assertEqual(1, parsed['day'])
+        self.assertEqual(0, parsed['hour'])
+        self.assertEqual(0, parsed['minute'])
+        self.assertEqual(0, parsed['second'])
+        self.assertEqual(0, parsed['subsecond'])
+        self.assertEqual(None, parsed['offset'])
+
+        text = '201205'
+
+        parsed = Parser().parse(text)
+        self.assertEqual(2012, parsed['year'])
+        self.assertEqual(5, parsed['month'])
+        self.assertEqual(1, parsed['day'])
+        self.assertEqual(0, parsed['hour'])
+        self.assertEqual(0, parsed['minute'])
+        self.assertEqual(0, parsed['second'])
+        self.assertEqual(0, parsed['subsecond'])
+        self.assertEqual(None, parsed['offset'])
+
+    def test_iso8601_datetime(self):
         text = '2016-10-01T14'
 
         parsed = Parser().parse(text)
@@ -231,6 +280,98 @@ class ParserTest(AbstractTestCase):
         self.assertEqual(0, parsed['second'])
         self.assertEqual(400000000, parsed['subsecond'])
         self.assertEqual(19800, parsed['offset'])
+
+        text = '2008-09-03T20:56:35.450686+01'
+
+        parsed = Parser().parse(text)
+        self.assertEqual(2008, parsed['year'])
+        self.assertEqual(9, parsed['month'])
+        self.assertEqual(3, parsed['day'])
+        self.assertEqual(20, parsed['hour'])
+        self.assertEqual(56, parsed['minute'])
+        self.assertEqual(35, parsed['second'])
+        self.assertEqual(450686000, parsed['subsecond'])
+        self.assertEqual(3600, parsed['offset'])
+
+    def test_iso8601_week_number(self):
+        text = '2012-W05'
+
+        parsed = Parser().parse(text)
+        self.assertEqual(2012, parsed['year'])
+        self.assertEqual(1, parsed['month'])
+        self.assertEqual(30, parsed['day'])
+        self.assertEqual(0, parsed['hour'])
+        self.assertEqual(0, parsed['minute'])
+        self.assertEqual(0, parsed['second'])
+        self.assertEqual(0, parsed['subsecond'])
+        self.assertEqual(None, parsed['offset'])
+
+        text = '2012W05'
+
+        parsed = Parser().parse(text)
+        self.assertEqual(2012, parsed['year'])
+        self.assertEqual(1, parsed['month'])
+        self.assertEqual(30, parsed['day'])
+        self.assertEqual(0, parsed['hour'])
+        self.assertEqual(0, parsed['minute'])
+        self.assertEqual(0, parsed['second'])
+        self.assertEqual(0, parsed['subsecond'])
+        self.assertEqual(None, parsed['offset'])
+
+        text = '2012-W05-5'
+
+        parsed = Parser().parse(text)
+        self.assertEqual(2012, parsed['year'])
+        self.assertEqual(2, parsed['month'])
+        self.assertEqual(3, parsed['day'])
+        self.assertEqual(0, parsed['hour'])
+        self.assertEqual(0, parsed['minute'])
+        self.assertEqual(0, parsed['second'])
+        self.assertEqual(0, parsed['subsecond'])
+        self.assertEqual(None, parsed['offset'])
+
+        text = '2012W055'
+
+        parsed = Parser().parse(text)
+        self.assertEqual(2012, parsed['year'])
+        self.assertEqual(2, parsed['month'])
+        self.assertEqual(3, parsed['day'])
+        self.assertEqual(0, parsed['hour'])
+        self.assertEqual(0, parsed['minute'])
+        self.assertEqual(0, parsed['second'])
+        self.assertEqual(0, parsed['subsecond'])
+        self.assertEqual(None, parsed['offset'])
+
+    def test_iso8601_ordinal(self):
+        text = '2012-007'
+
+        parsed = Parser().parse(text)
+        self.assertEqual(2012, parsed['year'])
+        self.assertEqual(1, parsed['month'])
+        self.assertEqual(7, parsed['day'])
+        self.assertEqual(0, parsed['hour'])
+        self.assertEqual(0, parsed['minute'])
+        self.assertEqual(0, parsed['second'])
+        self.assertEqual(0, parsed['subsecond'])
+        self.assertEqual(None, parsed['offset'])
+
+        text = '2012007'
+
+        parsed = Parser().parse(text)
+        self.assertEqual(2012, parsed['year'])
+        self.assertEqual(1, parsed['month'])
+        self.assertEqual(7, parsed['day'])
+        self.assertEqual(0, parsed['hour'])
+        self.assertEqual(0, parsed['minute'])
+        self.assertEqual(0, parsed['second'])
+        self.assertEqual(0, parsed['subsecond'])
+        self.assertEqual(None, parsed['offset'])
+
+    def test_iso8601_ordinal_invalid(self):
+        text = '2012-007-05'
+
+        self.assertRaises(ParserError, Parser().parse, text)
+
 
     def test_invalid(self):
         text = '201610T'